requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
pypdf
langchain-core
langchain-chroma
langchain-google-genai
sentence_transformers
langchain-huggingface
fastapi
uvicorn[standard]
python-multipart
protobuf>=3.20.0,<4.25.0